requirements.txt
absl-py==2.1.0
accelerate==0.27.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.0
controlnet_aux==0.0.7
cycler==0.12.1
diffusers==0.26.3
docutils==0.20.1
einops==0.7.0
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.49.0
fsspec==2024.2.0
huggingface-hub==0.21.4
idna==3.6
imageio==2.34.0
importlib_metadata==7.0.2
jaraco.classes==3.3.1
jax==0.4.25
jaxlib==0.4.25
Jinja2==3.1.3
keyring==24.3.1
kiwisolver==1.4.5
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mediapipe==0.10.11
ml-dtypes==0.3.2
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
nh3==0.2.15
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
openvino==2024.0.0
openvino-telemetry==2023.2.1
opt-einsum==3.3.0
packaging==24.0
pillow==10.2.0
pkginfo==1.10.0
protobuf==3.20.3
psutil==5.9.8
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
readme_renderer==43.0
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
safetensors==0.4.2
scikit-image==0.22.0
scipy==1.12.0
six==1.16.0
sounddevice==0.4.6
sympy==1.12
tifffile==2024.2.12
timm==0.9.16
tokenizers==0.15.2
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
transformers==4.38.2
twine==5.0.0
typing_extensions==4.10.0
urllib3==2.2.1
zipp==3.17.0
